pyproject.toml
[project]
name = "restaurant-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.12",
    "django-debug-toolbar>=6.2.0",
    "django-model-utils>=5.0.0",
    "djangorestframework>=3.16.1",
]
